# `ClaimPreviewTile` state-map optimizations