--- a/ui/component/claimPreviewTile/index.js
+++ b/ui/component/claimPreviewTile/index.js
@@ -1,19 +1,18 @@
 import { connect } from 'react-redux';
 import {
   makeSelectClaimForUri,
-  makeSelectIsUriResolving,
+  selectIsUriResolving,
   getThumbnailFromClaim,
-  makeSelectTitleForUri,
-  makeSelectChannelForClaimUri,
-  makeSelectClaimIsNsfw,
-  makeSelectClaimIsStreamPlaceholder,
+  selectTitleForUri,
+  isStreamPlaceholderClaim,
   selectDateForUri,
 } from 'redux/selectors/claims';
 import { doFileGet } from 'redux/actions/file';
 import { doResolveUri } from 'redux/actions/claims';
-import { makeSelectViewCountForUri, selectBanStateForUri } from 'lbryinc';
-import { makeSelectIsActiveLivestream } from 'redux/selectors/livestream';
+import { selectViewCountForUri, selectBanStateForUri } from 'lbryinc';
+import { selectIsActiveLivestreamForUri } from 'redux/selectors/livestream';
 import { selectShowMatureContent } from 'redux/selectors/settings';
+import { isClaimNsfw } from 'util/claim';
 import ClaimPreviewTile from './view';
 import formatMediaDuration from 'util/formatMediaDuration';
 
@@ -21,21 +20,21 @@ const select = (state, props) => {
   const claim = props.uri && makeSelectClaimForUri(props.uri)(state);
   const media = claim && claim.value && (claim.value.video || claim.value.audio);
   const mediaDuration = media && media.duration && formatMediaDuration(media.duration, { screenReader: true });
+  const isLivestream = isStreamPlaceholderClaim(claim);
 
   return {
     claim,
     mediaDuration,
     date: props.uri && selectDateForUri(state, props.uri),
-    channel: props.uri && makeSelectChannelForClaimUri(props.uri)(state),
-    isResolvingUri: props.uri && makeSelectIsUriResolving(props.uri)(state),
+    isResolvingUri: props.uri && selectIsUriResolving(state, props.uri),
     thumbnail: getThumbnailFromClaim(claim),
-    title: props.uri && makeSelectTitleForUri(props.uri)(state),
+    title: props.uri && selectTitleForUri(state, props.uri),
     banState: selectBanStateForUri(state, props.uri),
     showMature: selectShowMatureContent(state),
-    isMature: makeSelectClaimIsNsfw(props.uri)(state),
-    isLivestream: makeSelectClaimIsStreamPlaceholder(props.uri)(state),
-    isLivestreamActive: makeSelectIsActiveLivestream(props.uri)(state),
-    viewCount: makeSelectViewCountForUri(props.uri)(state),
+    isMature: claim ? isClaimNsfw(claim) : false,
+    isLivestream,
+    isLivestreamActive: isLivestream && selectIsActiveLivestreamForUri(state, props.uri),
+    viewCount: selectViewCountForUri(state, props.uri),
   };
 };