--- a/ui/component/claimPreviewTitle/index.js
+++ b/ui/component/claimPreviewTitle/index.js
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux';
-import { makeSelectClaimForUri, makeSelectTitleForUri } from 'redux/selectors/claims';
+import { makeSelectClaimForUri, selectTitleForUri } from 'redux/selectors/claims';
 import ClaimPreviewTitle from './view';
 
 const select = (state, props) => ({
   claim: makeSelectClaimForUri(props.uri)(state),
-  title: makeSelectTitleForUri(props.uri)(state),
+  title: selectTitleForUri(state, props.uri),
 });
 
 export default connect(select)(ClaimPreviewTitle);