--- a/ui/component/claimType/index.js
+++ b/ui/component/claimType/index.js
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux';
-import { makeSelectClaimForUri, makeSelectClaimIsStreamPlaceholder } from 'redux/selectors/claims';
+import { makeSelectClaimForUri, selectIsStreamPlaceholderForUri } from 'redux/selectors/claims';
 import FileType from './view';
 
 const select = (state, props) => ({
   claim: makeSelectClaimForUri(props.uri)(state),
-  isLivestream: makeSelectClaimIsStreamPlaceholder(props.uri)(state),
+  isLivestream: selectIsStreamPlaceholderForUri(state, props.uri),
 });
 
 export default connect(select)(FileType);