--- a/ui/component/fileTitleSection/index.js
+++ b/ui/component/fileTitleSection/index.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import { doFetchSubCount, makeSelectSubCountForUri } from 'lbryinc';
-import { makeSelectTitleForUri, makeSelectClaimForUri } from 'redux/selectors/claims';
+import { doFetchSubCount, selectSubCountForUri } from 'lbryinc';
+import { selectTitleForUri, makeSelectClaimForUri } from 'redux/selectors/claims';
 import { makeSelectInsufficientCreditsForUri } from 'redux/selectors/content';
 import { makeSelectViewersForId } from 'redux/selectors/livestream';
 import FileTitleSection from './view';
@@ -10,12 +10,12 @@ const select = (state, props) => {
   const viewers = claim && makeSelectViewersForId(claim.claim_id)(state);
   const channelClaimId = claim && claim.signing_channel ? claim.signing_channel.claim_id : undefined;
   const channelUri = claim && claim.signing_channel ? claim.signing_channel.canonical_url : undefined;
-  const subCount = channelUri && makeSelectSubCountForUri(channelUri)(state);
+  const subCount = channelUri && selectSubCountForUri(state, channelUri);
 
   return {
     viewers,
     isInsufficientCredits: makeSelectInsufficientCreditsForUri(props.uri)(state),
-    title: makeSelectTitleForUri(props.uri)(state),
+    title: selectTitleForUri(state, props.uri),
     channelClaimId,
     subCount,
   };