--- a/ui/component/fileType/index.js
+++ b/ui/component/fileType/index.js
@@ -1,11 +1,11 @@
 import { connect } from 'react-redux';
 import { makeSelectMediaTypeForUri } from 'redux/selectors/file_info';
-import { makeSelectClaimIsStreamPlaceholder } from 'redux/selectors/claims';
+import { selectIsStreamPlaceholderForUri } from 'redux/selectors/claims';
 import FileType from './view';
 
 const select = (state, props) => ({
   mediaType: makeSelectMediaTypeForUri(props.uri)(state),
-  isLivestream: makeSelectClaimIsStreamPlaceholder(props.uri)(state),
+  isLivestream: selectIsStreamPlaceholderForUri(state, props.uri),
 });
 
 export default connect(select)(FileType);