--- a/ui/component/fileViewerEmbeddedTitle/index.js
+++ b/ui/component/fileViewerEmbeddedTitle/index.js
@@ -1,11 +1,11 @@
 import { connect } from 'react-redux';
 import fileViewerEmbeddedTitle from './view';
-import { makeSelectTagInClaimOrChannelForUri, makeSelectTitleForUri } from 'redux/selectors/claims';
+import { makeSelectTagInClaimOrChannelForUri, selectTitleForUri } from 'redux/selectors/claims';
 import { PREFERENCE_EMBED } from 'constants/tags';
 
 export default connect((state, props) => {
   return {
-    title: makeSelectTitleForUri(props.uri)(state),
+    title: selectTitleForUri(state, props.uri),
     preferEmbed: makeSelectTagInClaimOrChannelForUri(props.uri, PREFERENCE_EMBED)(state),
   };
 })(fileViewerEmbeddedTitle);