--- a/ui/component/previewLink/index.js
+++ b/ui/component/previewLink/index.js
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux';
 import {
   selectClaimIsMine,
-  makeSelectTitleForUri,
+  selectTitleForUri,
   getThumbnailFromClaim,
   selectClaimForUri,
-  makeSelectIsUriResolving,
+  selectIsUriResolving,
   makeSelectMetadataItemForUri,
 } from 'redux/selectors/claims';
 import { doResolveUri } from 'redux/actions/claims';
@@ -17,11 +17,11 @@ const select = (state, props) => {
   return {
     uri: props.uri,
     claim,
-    title: makeSelectTitleForUri(props.uri)(state),
+    title: selectTitleForUri(state, props.uri),
     thumbnail: getThumbnailFromClaim(claim),
     description: makeSelectMetadataItemForUri(props.uri, 'description')(state),
     channelIsMine: selectClaimIsMine(state, claim),
-    isResolvingUri: makeSelectIsUriResolving(props.uri)(state),
+    isResolvingUri: selectIsUriResolving(state, props.uri),
     blackListedOutpoints: selectBlackListedOutpoints(state),
   };
 };