--- a/ui/component/uriIndicator/index.js
+++ b/ui/component/uriIndicator/index.js
@@ -1,12 +1,12 @@
 import { connect } from 'react-redux';
 import { normalizeURI } from 'util/lbryURI';
 import { doResolveUri } from 'redux/actions/claims';
-import { makeSelectIsUriResolving, makeSelectClaimForUri } from 'redux/selectors/claims';
+import { selectIsUriResolving, makeSelectClaimForUri } from 'redux/selectors/claims';
 import UriIndicator from './view';
 
 const select = (state, props) => ({
   claim: makeSelectClaimForUri(props.uri)(state),
-  isResolvingUri: makeSelectIsUriResolving(props.uri)(state),
+  isResolvingUri: selectIsUriResolving(state, props.uri),
   uri: normalizeURI(props.uri),
 });