## Issue
Lots of time is spent mapping the state to props for this component (since there are lots of tiles).

## Changes
Using this component as a starting point, went through the selectors and made the usual cleanup/fixes:
- Move away from the wrong `makeSelect*` model, which creates a new selector on every call instead of actually re-using the cached version.
- Do proper caching for multi-param selectors using `re-reselect`.
- Don't cache simple functions or direct access to states.

## Results
Here's a small sample of time saved when scrolling the homepage.  The first `select` in the list refers to `ClaimPreviewTile`, and the total time dropped from 250ms to 50ms, and the total mapping time for all components dropped 50% (580ms --> 220ms), since the selectors are shared by many components.

<img src="https://user-images.githubusercontent.com/64950861/141937920-1906cc03-907c-44b7-9d70-11faa611540d.png" width="400">      -      <img src="https://user-images.githubusercontent.com/64950861/141937763-220b29ac-849b-4519-9dbc-ac45a7cbf6e5.png" width="400">